refactor(web): フィーチャーディレクトリ配下の構成を整理
- 各フィーチャーの UI を components/ サブディレクトリに集約
- diary 配下の冗長な diary- prefix を除去し、diary-shared を daily-summary に改名
- features 直下にあった共通基盤を移動（create-swr-fetcher→infrastructure、汎用 validation→lib）
- features 内外の相対 import を @/* 絶対 import に統一
