+++ application/packages/web/src/client/infrastructure/create-swr-fetcher.test.ts
