+++ application/packages/web/src/client/routes/login/page.tsx
@@ -1,9 +1,9 @@
+import { AuthenticateForm } from '@/client/features/authenticate/components/authenticate-form'
+import { AuthenticateErrors } from '@/client/features/authenticate/validation'
 import { Card, CardContent, CardFooter, CardHeader, CardTitle } from '../../components/shadcn/card'
 import Footer from '../../components/ui/footer'
 import LandingHeader from '../../components/ui/landing-header'
 import { AnchorLink } from '../../components/ui/link'
-import { AuthenticateForm } from '../../features/authenticate/authenticate-form'
-import { AuthenticateErrors } from '../../features/authenticate/validation'
 
 interface Props {
   isSubmitting: boolean
