+++ application/packages/web/src/client/routes/signup/page.tsx
@@ -1,3 +1,5 @@
+import { AuthenticateForm } from '@/client/features/authenticate/components/authenticate-form'
+import { AuthenticateErrors } from '@/client/features/authenticate/validation'
 import {
   Card,
   CardContent,
@@ -9,8 +11,6 @@ import {
 import Footer from '../../components/ui/footer'
 import LandingHeader from '../../components/ui/landing-header'
 import { AnchorLink } from '../../components/ui/link'
-import { AuthenticateForm } from '../../features/authenticate/authenticate-form'
-import { AuthenticateErrors } from '../../features/authenticate/validation'
 
 interface Props {
   isSubmitting: boolean
