+++ application/packages/web/src/client/routes/signup/route.tsx
@@ -14,7 +14,7 @@ import {
 import {
   type AuthenticateErrors,
   validateAuthenticateForm,
-} from '../../features/authenticate/validation'
+} from '@/client/features/authenticate/validation'
 import SignupPage from './page'
 
 interface SignupActionData {
